workload/roachmart: relocate from pkg/ccl/workloadccl/roachmartccl
Epic: none
Release note: None
