+++ pkg/BUILD.bazel
@@ -1065,7 +1065,6 @@ GO_TARGETS = [
     "//pkg/ccl/workloadccl/allccl:allccl",
     "//pkg/ccl/workloadccl/allccl:allccl_test",
     "//pkg/ccl/workloadccl/cliccl:cliccl",
-    "//pkg/ccl/workloadccl/roachmartccl:roachmartccl",
     "//pkg/ccl:ccl",
     "//pkg/changefeed/changefeedpb:changefeedpb",
     "//pkg/changefeed/changefeedpb:checkpointpb",
@@ -3161,6 +3160,7 @@ GO_TARGETS = [
     "//pkg/workload/queue:queue",
     "//pkg/workload/rand:rand",
     "//pkg/workload/rand:rand_test",
+    "//pkg/workload/roachmart:roachmart",
     "//pkg/workload/schemachange:schemachange",
     "//pkg/workload/schemachange:schemachange_test",
     "//pkg/workload/sqlsmith:sqlsmith",
